@@ -1,8 +1,9 @@
 // @ts-check
 // Note: type annotations allow type checking and IDEs autocompletion
 
-const lightCodeTheme = require('prism-react-renderer/themes/github');
-const darkCodeTheme = require('prism-react-renderer/themes/dracula');
+import {themes as prismThemes} from 'prism-react-renderer';
+
+import "./plugins/docusaurus-plugin-svgfix/svgfix.ts";
 
 /** @type {import('@docusaurus/types').Config} */
 const config = {
@@ -11,7 +12,7 @@ const config = {
   favicon: 'img/favicon.ico',
 
   // Set the production url of your site here
-  url: 'https://docs.dodao.dev',
+  url: 'https://docs.dodao.dev/',
   // Set the /<baseUrl>/ pathname under which your site is served
   // For GitHub pages deployment, it is often '/<projectName>/'
   baseUrl: '/',
@@ -24,9 +25,9 @@ const config = {
   onBrokenLinks: 'throw',
   onBrokenMarkdownLinks: 'warn',
 
-  // Even if you don't use internalization, you can use this field to set useful
-  // metadata like html lang. For example, if your site is Chinese, you may want
-  // to replace "en" with "zh-Hans".
+  // Even if you don't use internationalization, you can use this field to set
+  // useful metadata like html lang. For example, if your site is Chinese, you
+  // may want to replace "en" with "zh-Hans".
   i18n: {
     defaultLocale: 'en',
     locales: ['en'],
@@ -38,7 +39,7 @@ const config = {
       /** @type {import('@docusaurus/preset-classic').Options} */
       ({
         docs: {
-          sidebarPath: require.resolve('./sidebars.js'),
+          sidebarPath: './sidebars.js',
           // Please change this to your repo.
           // Remove this to remove the "edit this page" links.
           editUrl:
@@ -52,7 +53,7 @@ const config = {
             'https://github.com/facebook/docusaurus/tree/main/packages/create-docusaurus/templates/shared/',
         },
         theme: {
-          customCss: require.resolve('./src/css/custom.css'),
+          customCss: './src/css/custom.css',
         },
       }),
     ],
@@ -71,8 +72,8 @@ const config = {
         },
         items: [
           {
-            type: 'doc',
-            docId: 'intro',
+            type: 'docSidebar',
+            sidebarId: 'tutorialSidebar',
             position: 'left',
             label: 'Docs',
           },
@@ -84,14 +85,6 @@ const config = {
           },
         ],
       },
-      // announcementBar: {
-      //   id: 'support_us',
-      //   content:
-      //     'We are looking to revamp our docs, please fill <a target="_blank" rel="noopener noreferrer" href="#">this survey</a>',
-      //   backgroundColor: '#fafbfc',
-      //   textColor: '#091E42',
-      //   isCloseable: false,
-      // },
       footer: {
         style: 'dark',
         links: [
@@ -144,10 +137,6 @@ const config = {
           {
             title: 'More',
             items: [
-              // {
-              //   label: 'Blog',
-              //   to: '/blog',
-              // },
               {
                 label: 'GitHub',
                 href: 'https://github.com/devopsdao',
@@ -162,28 +151,14 @@ const config = {
         copyright: `${new Date().getFullYear()} Dodao.dev . Built with Docusaurus.`,
       },
       prism: {
-        theme: lightCodeTheme,
-        darkTheme: darkCodeTheme,
+        theme: prismThemes.github,
+        darkTheme: prismThemes.dracula,
       },
     }),
 
-    plugins: [
-      function svgFix() {
-        return {
-          name: 'svg-fix',
-          configureWebpack(config) {
-            const svgRuleIndex = config.module.rules.findIndex((r) => r.test.test('file.svg'))
-            const svgrConfigIndex = config.module.rules[svgRuleIndex].oneOf.findIndex((r) => {
-              if (!Array.isArray(r.use) || r.use.length === 0) return false
-              return r.use[0].loader.indexOf('@svgr/webpack') !== -1
-            })
-            if (svgRuleIndex === -1 || svgrConfigIndex === -1) return
-    
-            config.module.rules[svgRuleIndex].oneOf[svgrConfigIndex].use[0].options.svgoConfig.plugins[0].params.overrides.cleanupIDs = false
-          }
-        }
-      }
-    ]
+    // plugins: [
+    //   "./plugins/docusaurus-plugin-svgfix/index.ts"
+    // ]
 };
 
-module.exports = config;
+export default config;
