@@ -1,5 +1,5 @@
-import React from 'react';
 import clsx from 'clsx';
+import Heading from '@theme/Heading';
 import styles from './styles.module.css';
 
 const FeatureList = [
@@ -39,7 +39,7 @@ function Feature({Svg, title, description, url}) {
   return (
     <div className={clsx('col col--4')}>
       <div className="text--center">
-      <a href={url}><Svg className={styles.featureSvg} role="img" /></a>
+        <Svg className={styles.featureSvg} role="img" />
       </div>
       <div className="text--center padding-horiz--md">
         <h3><a href={url}>{title}</a></h3>
